requirements.txt
notebook
python-jobspy
kafka-python
psycopg2-binary
sqlalchemy
pandas